pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "family-photo"
version = "0.1.0"
description = "CLI tools for processing scanned family photo album pages."
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.0",
  "opencv-python-headless>=4.10.0",
  "Pillow>=10.0.0",
  "pytesseract>=0.3.13",
  "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
handwriting-ocr = [
  "torch>=2.7.0",
  "transformers>=4.52.0",
]

[project.scripts]
family-photo = "cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
